Move HTML parsing to a separate file.
